+++ dask_geomodeling/tests/test_raster.py
@@ -1102,6 +1102,24 @@ class TestGroup(TestCombine, unittest.TestCase):
         data = view.get_data(mode="vals", width=1, height=1, **request)
         self.assertEqual(data["values"].tolist(), [[[view.fillvalue]]])
 
+    def test_only_gap_no_request_period_overlap(self):
+        view = self.klass(self.storage1)
+        request = dict(
+            start=Datetime(2000, 1, 1, 0, 20),  # after period
+            stop=Datetime(2000, 1, 1, 0, 25),  # even more after period
+        )
+        _requests = view.get_sources_and_requests(mode="meta", **request)
+        self.assertEqual(_requests[0][0]["combine_mode"], "simple")
+
+        time = view.get_data(mode="time", **request)
+        assert time is None
+
+        meta = view.get_data(mode="meta", **request)
+        assert meta is None
+
+        data = view.get_data(mode="vals", width=1, height=1, **request)
+        assert data is None
+
 
 class TestSnap(unittest.TestCase):
     klass = raster.Snap